moved set_camera_zoom and do_set_servo to commander
